refactor(infra): renames <Select/> stories, component file
